Adds unit test for ConcurrentTestRunner
