+++ geode-concurrency-test/build.gradle
@@ -14,11 +14,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-apply from: "${project.projectDir}/../gradle/publish.gradle"
 
+apply from: "${project.projectDir}/../gradle/publish.gradle"
 
 dependencies {
   compile(platform(project(':boms:geode-all-bom')))
   compile('junit:junit')
   compile('org.apache.logging.log4j:log4j-api')
+  testCompile('org.assertj:assertj-core')
 }
