+++ geode-core/src/main/java/org/apache/geode/cache/client/internal/pooling/ConnectionAccounting.java
@@ -26,7 +26,7 @@ public class ConnectionAccounting {
   private final int min;
   private final int max;
   private final AtomicInteger count = new AtomicInteger();
-//   private final NotAtomicInteger count = new NotAtomicInteger();
+  // private final NotAtomicInteger count = new NotAtomicInteger();
 
   public ConnectionAccounting(int min, int max) {
     this.min = min;
